requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
defusedxml==0.7.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
httpie==3.2.4
idna==3.13
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
pillow==12.2.0
Pygments==2.20.0
PyJWT==2.12.1
PySocks==1.7.1
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
setuptools==82.0.1
sqlparse==0.5.5
urllib3==2.6.3
